Fix more paths, 79 passes
